USER Modify the file update method to replace the current content with the new one
